pyproject.toml
[tool.poetry]
name = "lasfera-app"
version = "0.1.0"
description = ""
authors = ["Jason Heppler <jheppler@gmu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
django-allauth = "^0.61.1"
django-environ = "^0.11.2"
pylint-django = "^2.5.5"
pillow = "^10.2.0"
psycopg2 = "^2.9.9"
django-admin-interface = "^0.28.5"
pandas = "^2.2.1"
numpy = "^1.26.4"
openpyxl = "^3.1.2"
django-prose = "^2.0.0"
geopy = "^2.4.1"
django-nested-admin = "^4.0.2"
django-dbml = "^0.7.0"
beautifulsoup4 = "^4.12.3"
django-prose-editor = "^0.3.4"
django-tinymce = "^4.0.0"
django-import-export = "^4.0.3"
django-tailwind = "^3.8.0"
djangorestframework = "^3.15.2"
fontawesomefree = "^6.6.0"


[tool.poetry.group.development.dependencies]
django-tailwind = {extras = ["reload"], version = "^3.8.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
